// src/interfaces/customer.ts
// Generated by https://quicktype.io

export interface Customer {
  id: number;
  firstName: string;
  lastName: string;
  phoneNumber: string;
  birthDate: string;
  createdAt: string;
  updatedAt: string;
  lastLogin: null;
  receiveAds: boolean;
  user: User;
}

export interface User {
  id: number;
  email: string;
  role: string;
  createdAt: string;
  updatedAt: string;
  lastLogin: null;
  isEmailConfirmed: boolean;
}